LayoutTests/platform/chromium-win/fullscreen/full-screen-placeholder-expected.txt

This layout test checks that the offset positions of the blue and green divs does not change when the red div enters full-screen mode. Press go full-screen to begin.
One
Two
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLDivElement]') OK
EXPECTED (one.offsetLeft == '68') OK
EXPECTED (one.offsetTop == '54') OK
EXPECTED (two.offsetLeft == '8') OK
EXPECTED (two.offsetTop == '114') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == 'null') OK
EXPECTED (one.offsetLeft == '68') OK
EXPECTED (one.offsetTop == '54') OK
EXPECTED (two.offsetLeft == '8') OK
EXPECTED (two.offsetTop == '114') OK
END OF TEST
